ppath object decomposed from pcompat
